React type dependencies into package. Inferno definitions.
